Unify transformer_lm_megatron and transformer_lm
