+++ metaseq/models/transformer_lm.py
@@ -3,10 +3,28 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import torch
+import torch.nn as nn
+from metaseq.models.transformer_decoder import (
+    ModelParallelTransformerDecoder,
+)
+
+
+try:
+    from megatron.mpu import VocabParallelEmbedding
+
+    has_megatron_submodule = True
+except (ImportError, ModuleNotFoundError):
+    has_megatron_submodule = False
+
+
+DEFAULT_MAX_TARGET_POSITIONS = 1024
 
 import logging
 
 from dataclasses import dataclass, field
+
+from metaseq.dataclass import ChoiceEnum, MetaseqDataclass
 from typing import Optional
 
 from omegaconf import II
@@ -14,7 +32,6 @@ from omegaconf import II
 from metaseq.dataclass.constants import ATTN_CHOICES, UNSPECIFIED_DOC_SEP
 
 from metaseq import utils
-from metaseq.dataclass import ChoiceEnum, MetaseqDataclass
 from metaseq.models import (
     LanguageModel,
     register_model,
@@ -27,10 +44,7 @@ from metaseq.models.transformer_decoder import (
 from metaseq.modules.embedding import Embedding
 from metaseq.modules.activation_functions import get_available_activation_fns
 
-DEFAULT_MAX_TARGET_POSITIONS = 1024
-
 logger = logging.getLogger(__name__)
-
 
 @dataclass
 class TransformerLanguageModelConfig(MetaseqDataclass):
@@ -188,7 +202,7 @@ class TransformerLanguageModelConfig(MetaseqDataclass):
     model_parallel_size: int = II("common.model_parallel_size")
 
 
-@register_model("transformer_lm", dataclass=TransformerLanguageModelConfig)
+@register_model("base_transformer_lm", dataclass=TransformerLanguageModelConfig)
 class TransformerLanguageModel(LanguageModel):
     def __init__(self, decoder):
         super().__init__(decoder)
@@ -225,30 +239,138 @@ class TransformerLanguageModel(LanguageModel):
         )
 
 
+@register_model("model_parallel_transformer_lm")
+class ModelParallelTransformerLanguageModel(LanguageModel):
+    @classmethod
+    def build_model(cls, args, task):
+        """Build a new model instance."""
+        if not has_megatron_submodule:
+            raise ImportError(
+                "\n\nPlease install megatron using the setup instructions!"
+            )
+
+        # make sure all arguments are present in older models
+        base_lm_architecture(args)
+
+        task.source_dictionary.pad_to_multiple_(8)
+        task.target_dictionary.pad_to_multiple_(8)
+
+        # task.source_dictionary.pad_to_multiple_(args.model_parallel_size * 8)
+        # task.target_dictionary.pad_to_multiple_(args.model_parallel_size * 8)
+
+        if getattr(args, "max_target_positions", None) is None:
+            args.max_target_positions = getattr(
+                args, "tokens_per_sample", DEFAULT_MAX_TARGET_POSITIONS
+            )
+
+        embed_tokens = cls.build_embedding(
+            args, task.source_dictionary, args.decoder_embed_dim
+        )
+        assert getattr(
+            args, "use_sharded_state", False
+        ), "Use sharded state must be True for tensor parallel, otherwise model saving and loaded might be broken"
+
+        if getattr(args, "sequence_parallel", False):
+            assert (
+                getattr(args, "model_parallel_size", 1) > 1
+            ), "--sequence-parallel only works when --model-parallel-size is greater than 1"
+            assert (
+                getattr(args, "dropout", 0.0) == 0.0
+            ), "havent yet tested if rng states are correct for dropout with seq_parallel"
+            assert (
+                getattr(args, "activation_fn", "gelu") == "gelu"
+                or getattr(args, "activation_fn", "gelu") == "relu"
+            ), "For now only supports gelu and relu"
+            assert not getattr(
+                args, "checkpoint_activations", False
+            ), "Cannot set --checkpoint-activations with sequence parallel."
+            assert not getattr(
+                args, "distribute_checkpointed_activations", False
+            ), "Cannot set --distribute-checkpointed-activations with sequence parallel."
+
+        decoder = ModelParallelTransformerDecoder(
+            args,
+            task.target_dictionary,
+            embed_tokens,
+        )
+        return cls(decoder)
+
+    @staticmethod
+    def add_args(parser):
+        TransformerLanguageModel.add_args(parser)
+
+    @classmethod
+    def build_embedding(cls, args, dictionary, embed_dim, path=None):
+        def _vocab_init(tensor, **kwargs):
+            std = embed_dim**-0.5
+            if getattr(args, "truncate_init", False):
+                nn.init.trunc_normal_(tensor, mean=0, std=std, a=-3 * std, b=3 * std)
+            else:
+                nn.init.normal_(tensor, mean=0, std=std)
+            nn.init.constant_(tensor[1], 0)
+
+        def _vocab_init_megatron(tensor, **kwargs):
+            std = getattr(args, "megatron_init_sigma", 0.006)
+            if getattr(args, "truncate_init", False):
+                nn.init.trunc_normal_(tensor, mean=0, std=std, a=-3 * std, b=3 * std)
+            else:
+                nn.init.normal_(tensor, mean=0, std=std)
+            nn.init.constant_(tensor[1], 0)
+
+        if getattr(args, "memory_efficient_fp16", False):
+            dtype = torch.bfloat16 if getattr(args, "bf16", False) else torch.half
+        else:
+            dtype = torch.float32
+
+        embed_tokens = VocabParallelEmbedding(
+            len(dictionary),
+            embed_dim,
+            dictionary.pad(),
+            init_method=_vocab_init_megatron
+            if getattr(args, "full_megatron_init", False)
+            else _vocab_init,
+            use_cpu_initialization=not getattr(
+                args, "tensor_parallel_init_model_on_gpu", False
+            ),
+            dtype=dtype,
+        )
+        return embed_tokens
+
+
 def base_lm_architecture(args):
+    args.activation_fn = getattr(args, "activation_fn", "relu")
     args.dropout = getattr(args, "dropout", 0.1)
     args.attention_dropout = getattr(args, "attention_dropout", 0.0)
-
     args.decoder_embed_dim = getattr(args, "decoder_embed_dim", 512)
     args.decoder_ffn_embed_dim = getattr(args, "decoder_ffn_embed_dim", 2048)
     args.decoder_layers = getattr(args, "decoder_layers", 6)
     args.decoder_attention_heads = getattr(args, "decoder_attention_heads", 8)
-    args.decoder_learned_pos = getattr(args, "decoder_learned_pos", False)
-    args.decoder_learned_sinusoidal = getattr(args, "decoder_learned_sinusoidal", False)
-    args.activation_fn = getattr(args, "activation_fn", "relu")
-
-    args.add_bos_token = getattr(args, "add_bos_token", False)
     args.share_decoder_input_output_embed = getattr(
         args, "share_decoder_input_output_embed", False
     )
+    args.decoder_learned_pos = getattr(args, "decoder_learned_pos", False)
+    args.decoder_learned_sinusoidal = getattr(args, "decoder_learned_sinusoidal", False)
     args.no_scale_embedding = getattr(args, "no_scale_embedding", False)
+    args.add_bos_token = getattr(args, "add_bos_token", False)
     args.checkpoint_activations = getattr(args, "checkpoint_activations", False)
     args.offload_activations = getattr(args, "offload_activations", False)
     if args.offload_activations:
         args.checkpoint_activations = True
 
 
-@register_model_architecture("transformer_lm", "transformer_lm_gpt")
+@register_model_architecture("model_parallel_transformer_lm", "transformer_lm_megatron")
+def transformer_lm_megatron(args):
+    args.decoder_embed_dim = getattr(args, "decoder_embed_dim", 3072)
+    args.decoder_ffn_embed_dim = getattr(args, "decoder_ffn_embed_dim", 3072 * 4)
+    args.decoder_layers = getattr(args, "decoder_layers", 72)
+    args.decoder_attention_heads = getattr(args, "decoder_attention_heads", 32)
+    args.dropout = getattr(args, "dropout", 0.1)
+    args.attention_dropout = getattr(args, "attention_dropout", 0.1)
+    args.activation_fn = getattr(args, "activation_fn", "gelu")
+    base_lm_architecture(args)
+
+
+@register_model_architecture("model_parallel_transformer_lm", "transformer_lm_gpt")
 def transformer_lm_gpt(args):
     args.decoder_embed_dim = getattr(args, "decoder_embed_dim", 768)
     args.decoder_ffn_embed_dim = getattr(args, "decoder_ffn_embed_dim", 3072)
@@ -260,7 +382,7 @@ def transformer_lm_gpt(args):
     base_lm_architecture(args)
 
 
-@register_model_architecture("transformer_lm", "transformer_lm_gpt2_tiny")
+@register_model_architecture("model_parallel_transformer_lm", "transformer_lm_gpt2_tiny")
 def transformer_lm_gpt2_tiny(args):
     args.decoder_embed_dim = getattr(args, "decoder_embed_dim", 64)
     args.decoder_ffn_embed_dim = getattr(args, "decoder_ffn_embed_dim", 64)
